hydra: fix fn_fullinit + spawn
The fullinit PMI command is used in both the pmi port path and PMI2.
With pmi port, we have to pass the correct pg to the callback or we'll
have no way locating the correct downstream in a multiple-pg scenario,
such as the case of spawn. Under PMI2, fullinit is used but the correct
downstream is available as the callback argument.

Instead of blindly matching downstream by rank, which won't work in a
spawned case, check whether the callback arg is an integer less than
1024; if true assume it is a pg_idx and set it to the pseudo downstream.
The pseudo downstream can be identified by checking its idx is -1, and
replaced with the correct downstream in fn_fullinit.