session: add strict_finalize info parameter.
Add non-standard info parameter strict_finailze for MPI Sessions.
The feature is disabled by default. If strict_finalize is enabled
MPI_Session_finalize will return an MPI_ERR_PENDING error if there
are still references on the session due to missing free for an MPI
object or a pending communication.
In addition, progress is triggered for the communication to allow
users to poll for success of MPI_Session_finalize with this option.
The goal of the strict_finalize parameter is to ensure that a
session has been completely cleaned up prior to finalization and
potential re-init.