config: remove --with-pmilib option, tune --with-pmi
Remove the (deprecated) configure option --with-pmilib.

The option --with-pmi is tuned to accept additional values embedded,
install, slurm-pmi2, oldcray-pmi2, and bgq. In addition, it accepts
as before pmi1, <path>, pmi2, pmix, slurm and oldcray.

Based on the specified option, the desired pmi[12x] library is
configured and the supported PM interface(s) is (are) enabled.

For now, using libpmix together with MPICHs libpmi is possible by
configuring with:
(--with-pmi=embedded || --with-pmi=install) && --with-pmix=<path>.
This way, all PM interfaces PMI1, PMI2, and PMIx are made available
in the MPI library.

Allow compile with hydra as long as either PMI1 or PMI2 are enabled.