mpir_pmi: use uthash in normalizing node ids
The previous code assumes the node id from process manager is within a
finite range. This may not be true. For example, a process manager may
use a bit pattern to represent node id, which can easily be a huge or
even negative integer. Thus, it is not safe to allocate a temp array of
size max_id+1. Use a hash instead.