nodemap: support node ids for dynamic processes
To support dynamic processes to use shared memory, we need match the new
process to the node_id of an existing process. For universal support, we
maintain a hash table for hostname to node_id mapping. Node ids for
comm_world is initialized at post init stage. New processes are assigned
with existing node_id (if matching hostname) or with new node_id as we
connect them. The latter will be supported in ch4 avtable management.