romio/gpfs: Add check against integer overflow
With f505565, the auxiliary function MY_Alltoallv was introduced, which
is used if MPI_Alltoallv_c is not available (i.e. in case of < MPI-4),
and if the displacesments are too large for integers.

In such a case, this function copies all that data into a new send buffer
before the common Alltoallv is called, but does not check whether this
can actually be covered in size by integer values.

This commit adds such a check and serves, since MPI_Alltoallv_c is
available with today's mpich versions, only for a potential backport.