romio/gpfs: Fix copy/paste error
To check whether the data buffers in case of the MY_Alltoall workaround
have actually to be moved, the type-casted sdispl[]/rdispl[] items are
to be compared -- but each with the entries of the same kind.

This commit fixes a suspected copy/paste bug in this procedure.

Since MY_Alltoall is only used if MPI_Alltoallv_c is not available (i.e.
in case of < MPI-4), this commit only serves as a fix for a potential
backport.