init: create built-in comms only for world model
Built-in comms MPI_COMM_WORLD and MPI_COMM_SELF are only
initialized for the world model, i.e., if MPI_Init/ MPI_
Finalize are used. In case only MPI Sessions are used, the
built-in comms are not initialized. Using MPI_COMM_WORLD or
MPI_COMM_SELF with the session model will raise errors as
these are MPI_COMM_NULL in this case.