pmi: check whether PMI2_keyval_t is missing
Cray PMI does not define PMI2_keyval_t. Check it in configure and
typedef INFO_TYPE PMI2_keyval_t if needed. This will allow mpich to compile using
Cray PMI. The PMI2_keyval_t is only used in PMI2_Job_Spawn. The name
publishing API always uses NULL for its info pointers. It is possible
that our INFO_TYPE is incompatible with Cray PMI's internal type, which
will break PMI2_Job_spawn. If so, this need be fixed in the future,
possibly on the Cray side.