mpir_pmix: util functions to manage info objects
With pmix_add_to_info, a key contained in an MPI info object
can be mapped to a PMIx key and the respective key-value pair
is inserted into a target info object using the PMIx key.

With mpi_to_pmix_keyvals, all key-value pairs contained in an
MPI Info object are transferred to a pmix_info_t object.