coll: add topology-aware collectives
This commit implements three types of topology-aware trees for dragonfly
network topologies.
The topology_aware tree builds a tree based on the dragonfly network
hierarchy of the actual nodes used by the ranks participating in the
collective.
The topology_aware_k tree is similar to the topology_aware tree, but it
controls the branching factor k.
The topology_wave tree builds a tree considering the latency/distance
between the nodes of the ranks that participate in the collective.

Co-authored-by: Xi Luo <xi.luo@intel.com>
Co-authored-by: Nikita Gusev <nikita.gusev@intel.com>
Co-authored-by: Todd Schwartz <todd.schwartz@intel.com>