configure: allow independent pmi options
Allow users to force --with-pm and --with-pmilib options. It allows
users to build hydra even with external pmi library or build embedded
libpmi concurrently with an external pmi library (e.g. libpmix). This
may not work due to conflicts, but it is at user's descretion.

Use may load 3rd party pmi libraries with:
   --with-pmi1=path   (libpmi.so)
   --with-pmi2=path   (libpmi2.so)
   --with-pmix=path   (libpmix.so)

As of this commit, following options are more or less independent:
   --with-pmi=default|pmi1|pmi2|pmix
   --with-pmilib=default|no|mpich|install
   --with-pm=default|no|gforker|remshell|hydra

The default option can be internally overwritten. with_pmilib=no will
not build internal(embedded) libpmi. with_pm=no will not build internal
pm.

A special option is --with-pmi=slurm. This is mainly because Slurm installs
header as #include "slurm/pmi.h".