mpir_pmix: fence with spawned processes and parent
For spawning with PMIx we need to ensure that spawned processes
read the preput values from the KVS (i.e., parent port string)
before the next spawn is launched in the parent process. A
PMIx_Fence with the processes in the spawned namespace + the
parent process ensures that. The spawned processes call PMIx_Fence
after they read the preput values from the KVS.