hydra/mpiexec: use utarray in parsing config file
When parsing config file, it first load tokens into a pre-allocated
array of size of HYD_NUM_TMP_STRINGS (1000). The hard limit is not safe
in case of very large number of arguments. Use utarray instead.

It seems the previous code also leaks the strdup'ed token strings. This
commit fixes that as well.