pmi: Add utility for getting PMI fd
All three client implementations use similar logic for getting a
connection to the PMI server. Unify under a single utility function. The
only difference is the PMI1 client needs an additional handshake with
the server if PMI_PORT is in use. This is indicated through an
additional output parameter that can be ignored in the PMI2 and PMIx
clients.