pmi: PMIU_read_cmd check fd for last_read
When it saves remaining buffer for next read, we need save the fd as
well so data from different fd does not mix. This is not the case on the
client side, but can arise on the server side. This commit makes the
utility robust in case we use the same utility on the server side.