pmi: Process "set" responses to "fullinit" in PMI1 client
When using PMI_PORT, the PMI1 client sends a "fullinit" command to the
server. The response to this command is an "initack" plus three "set"
commands for the size, rank, and debug values. The wire protocol layer
was processing these "set" messages internally, obfuscating them from
the client. Move this message processing into the client to help with
understanding.