request: set status MPI_ERROR fields in MPI_ERR_IN_STATUS
In MPIR_{Test,Wait}some, we only start to set the status MPI_ERROR field
when we see an error. But by then, the earlier status entries were left
MPI_ERROR field unset. This error is detected by
errors/pt2pt/errinstat{ts,ws} tests. Instead, we set all the status
MPI_ERROR field in a separate loop once we know we are returning
MPI_ERR_IN_STATUS.

Note that in MPIR_{Test,Wait}all, we always set MPI_ERROR fields
regardless whether we are returning MPI_ERR_IN_STATUS. This of course is
inconsistent to the strict interpretation of MPI standard (pt2pt
chapter, section "Return Status"), which says we should only set the
MPI_ERROR fields if we are returning MPI_ERR_IN_STATUS. However, I am
not sure this strict interpretation is correct or makes sense. The
intention to be give implementation flexibility so it don't need to
update MPI_ERROR field if it is returning an error anyway. The strict
interpretations appears to be countering this rationale.