ch4/ofi: refactor match BITS constants
MPIDI_OFI_PROTOCOL_BITS is defined in ofi_types.h, but the rest of the
bits are defined in ofi_capabilities.h. This allows inconsistency to
easily creep in. This is the case as in commit 32e5db5 when we modified
MPIDI_OFI_PROTOCOL_BITS without checking against *all* providers.
Refactor by defining bit patterns in ofi_types.h and use the predefined
constants in ofi_capabilities.h. We still can overwrite them, such is
the case for CXI provider, but at least the default will be consistent.