errhan: encode dynamic class in dynamic errorcodes
We used to use a single bit to mark an error code or an error class as
dynamic. But we cannot tell a dynamic error code with a dynamic class or
with a builtin error class. This is needed to properly reference count
error classes. Reuse the ERROR_FATAL_MASK bit for this purpose since
dynamic error code cannot be fatal.