errhan: always set ERROR_DYN_MASK in MPI_Add_error_code
When MPI_Add_error_code is called with a predefined class, the
ERROR_DYN_MASK is not set. This is a bug exist for a while. The old code
does fail the test especially at MPIX_Delete_error_code is because the
old MPIX_Delete_error_code implementation did not check ERROR_DYN_MASK
and assume it is always called with a dynamic error code.