ch3: skip win_free on allocate error for now
Ideally upon error in win_allocate_shared, we need free the window to
prevent resource leaks. However, we need be careful since there may have
unclean fields in the windown structure and the current window free is
causing double free error, likely due to trying to free an uninitialed
pointer.

Because ch3 is in legacy mode and this only affects resource leak under
an error condition, let's omit the win_free for now to prevent errors in
win_free.