ch4/mpidig: add comment on mpidig request status
Part of difficulty in understanding the code is due to the status macro
does not fully reflect the nature of the message. For example,
MPIDIG_REQ_BUSY is only relevant for unexpected eager message between
local data copy. It is confusing just looking at the "BUSY" label. Add a
comment block to clarify.