comm: fix intercomm split when root is not in
Intercomm split need exchange new_context id and we run collectives over
the old communicator. When root is not in, i.e. with color
MPI_UNDEFINED, it uses -1 for new context id, and resulting in
exchanging the wrong id for the processes that are in.

Rather than adding code to let root know the correct id, let root always
paticipating in generating the new id, then release the id afterward.