misc: rename MPIR_gpu_req to MPIR_async_req
MPIR_gpu_req is a union type for either a MPL_gpu_request or a
MPIR_Typerep_req, thus it is not just for gpu. Potentially this type can
be extended to include other internal async task handles. Thus we rename
it to MPIR_async_req.

We also establish the convention of naming the variable async_req.