mpl/shm: Add error checking for failed segment creation
In case the creation of the shared memory segment fails, check for and
return an error immediately. Otherwise the code would continue and crash
on a write to unallocated address. This was discovered when our test
machines ran out of space in /dev/shm, causing the lseek/write to fail
with ENOSPC.