ch4/ofi: move huge.remote_info in MPIDI_OFI_request_t
It is part of the recv path, thus need be in the recv struct.

The recv paths may switch paths depending on the actual protocols
used by sender, thus some of the fields from different paths need live
in the same struct and use the NULL sentinel to tell whether certain
path is in effect. Currently this include remote_info for huge_send
protocol, and pack_buffer for pack_recv. It's possible to have
huge_recv+pack_recv.