ch4/ofi: assert for unexpected pipeline data
There is no way to recover a pipelined message in the native ofi path if
the receiver didn't expect it. We could always save the necessary
information in all recv paths, but I am not sure we are willing to
accept the overhead, which harms the small message latencies.

Assert failure for now.

The proper pipeline implementation need to happen as active messages or
even at the MPIR-layer.