build: avoid makefile automatic variables in explicit rules
Automatic variables such as $< is not supported by non-gnumake unless
its under suffix rules. This fixes building on FreeBSD.

We are omitting $(top_builddir) since it will always be '.' and we
didn't use $(top_builddir) in other places bin_SCRIPTS.

Note: the previous version used $? which somehow worked. In this fix, we
are spelling the target out to make sure it will work universally.