test/coll: Use a single GPU per process in reduce.c
The reduce.c test used rank in comm to specify the GPU device used in
memory allocation. While iterating over multiple comms, the device could
change. This is an uncommon use-case, and one that is not yet supported
by UCX. Change the test to use world rank to specify the device. If
desired, we can add a multi-GPU test separately to verify that
capability.