mpl/gpu/ze: implement ipc cache size limit
Implement a size limit for the IPC mapped buffer cache with an LRU
replacement policy.

Prior to this addition, the cache for opened IPC handles (i.e.
containing mapped GPU buffers), could grow infinitely. If an IPC handle
is open, the original GPU buffer does not release resources when freed,
since the refcount is not 0. Thus, without a cache size limit, GPU are
not released until MPL_gpu_finalize when the cache is cleared.

With this change, a cache limit is enforced via an LRU replacement
policy. An opened IPC handle which is evicted is now closed so that GPU
resources can be freed during application runtime prior to finalize.