ch4/ofi: Apply FI_ORDER_SAS hint to rx_attr
According to discussion on the OFIWG Slack "...tx_attr at the sender
should align with the rx_attr at the receiver". The reason being that
the receiver may decide to reorder messages independent of the
sender. This was observed with certain providers when using FI_PEEK to
probe for messages when rx_attr did not set FI_ORDER_SAS.