test/io: fix i_noncontig_coll2.2
The test uses MPI_Gatherv to gather procnames into a `char **procname`
and uses `int *disp` to capture the difference between pointers. It
overflows on 64-bit systems when two malloc'ed pointers differ more than
INT_MAX. We fix it by using a single malloc'ed procname_buffer.

It is not clear why we didn't catch this issue in previous CI testing.