python/binding_c: Add streamsync directive
Extend the functionality of MPIX stream by allowing passing a stream
communication with local gpu stream to regular MPI functions. The
semantics is to run an implicit gpu stream synchronize before the MPI
operation. This is semantically serializing the MPI operation with the
GPU stream, albeit in a heavy handed way.