ch4/ofi: fix MPIDI_OFI_CALL_RETRY_AM for vci
The macro needs a vci input or it will call MPIDI_OFI_progress_do_queue
on the wrong vci and result in deadlock.

Somehow we miss catching this with psm3 provider testing. It is
triggered by am fallback in the rma testing. Apparently, only the rma
testing triggers the am postponing queue path.