ch4/ofi: add MPIR_CVAR_CH4_OFI_ENABLE_INJECT
The ofi injected messages may require explicit progress to kick the
message out. But this is difficult to control due to lack of request
handles. This is problemetic when user issue inject then immediately
dive into computation, or when they move on to a different dedicated vci
and neglect to progress the previous vci, which, by all indication,
has completed and doesn't require progress.

Add the cvar as a remedy.

We still believe this is a libfabric issue.