pmi: put value key to the end of cmd
Flux currently assumes the value key in a PMI_KVS_Get command will place
the value=... at the end of the wire command so that the value can
contain arbitrary chars including spaces and the `=`.  While we don't
want to make this the standard protocol, there is no reason that we
can't keep Flux working.

To keep consistency, we also exchange the order
of the "found" and "value" keys in the three PMI-v2 commands.