mpl/gpu: change MPL_gpu_device_handle_t to int
The abstraction of device id as an integer is a good abstraction above
MPL. The opaqueness of MPL_gpu_device_handle_t, on the other hand, makes
it useless, since the upperlayer can't do anything with it.

For ZE, simply expose the internal device id. For new GPU runtimes that
does support integer device ids, we can similarly do a map as in
mpl_gpu_ze.c.