mpir/pmi: fix use of PMIx_Load_topology
The "ptopo.source" from PMIx_Load_topology may not be "hwloc". Newer
version of pmix will append the version of hwloc, e.g. "hwloc:2.9.0".
Thus, we need to use strncmp instead of strcmp.

In case PMIx_Load_topology fail, we have the option of fallback. But we
may want to believe that PMIx_Load_topology should work and see its
error if it fails.