coll/csel: omit prunning on communicator size
Now that we may run collectives on subgroups, we can't pre-prune the
csel trees based on communicator size or topology since that may change for
subgroups.

I don't think the performance from the tree pruning is significant -- it
only saves a couple levels of tree decendence. But if we later decide
the efficiency from pruning is important, we can easily prune the trees at
subgroup level and save the pruned trees to the MPIR_Group structure.