coll: replace MPIR_ERR_COLL_CHECKANDCONT
Replace MPIR_ERR_COLL_CHECKANDCONT with MPIR_ERR_CHECK. Propagating
errors in collective does not work due to the complexity of collective
algorithms. For example, the error condition is not guaranteed to be
propagated to all processes. In addition, when there is random hardware
issue preventing the message to be delivered, trying to propagating
error only hides the error and result in hang anyway.