mpl/ze: fix a crash in implicit scaling mode at finalize
With implicit mode, two handles are returned from an extension function
zexMemGetIpcHandles, however, when return these handles, level zero
lacks of a corresponding version to handle two handles - Level zero
needs a new extension function for that. A temporary fix is not to call
putIpcHandle when it is implicit scaling mode.