ch4/ofi: remove the tinject path within send_normal
It seems the only reason for inject within send_normal is when syncflag
is on and we have to allocate an sreq. Since the benefit of inject send
is insignicant with synchronized send, let's simplify it away.

Lightweight send path decision now will solely be made in
MPIDI_OFI_send.