ch4/mpidig: refactor update request status at matching
The source and tag fields in MPIDIG_REQUEST and MPIDI_PART_REQUEST are
redundant to the MPI_SOURCE and MPI_TAG fields in the status, which needs
to be set anyway. Both the source and tag are confirmed upon receiving
the first active message thus they should be set then.

* remove MPIDIG_REQUEST(rreq, u.recv.{source,tag})and use
rreq->status.{MPI_SOURCE,MPI_TAG} instead.
* set or update the status fields at the time of receiving the first am
packet or at the time of matching.
* remove MPIR_TAG_MASK_ERROR_BITS in MPIDIG_match_request since we do
not set tag error bits in MPIR-layer or MPIDIG-layer (even if error-bits
are enabled).