binding/f08: preserve status%MPI_ERROR
Even though the status is marked as output in functions like MPI_Probe
and MPI_Recv, we expect the MPI_ERROR field to be preserved from input.
It is questionable whether this interpretation is right and we may
change the test mprobef08 in the future, for now, fix it by adding input
conversion of status.

This only affects Fortran binding when integer is different from C int.