pmi: increase MAX_TMP_BUF_SIZE to 4096
We occasionally get assertion errors like -

    pmi_wire.c:683: PMIU_cmd_output_v1: Assertion
`!PMIU_cmd_is_static(pmicmd)' failed.

This is due to PMI command size exceed MAX_TMP_BUF_SIZE and the upper
layer is not prepared for dynamic PMI command buffers. It's not clear
the exact issue, but I am guessing this is due to various limit in PMI1
such as PMI_MAXVALLEN is also set to 1024, which may result in combined
PMI command exceed 1024 limit. Thus, let's increase the internal buffer
to 4096 to have plenty of margin. Hopefully, this fixes the issue. If
the error continue to show up, then we need proper investigation.