ch4/ofi: Add NIC information to error output
On systems with multiple NICs, it could be helpful to know which NIC an
error was detected on in case there are hardware issues that need
investigating. Add NIC information to error checking macros. For now, we
report the default NIC used by each process. TODO: extend to support
multi-nic usage and take the device number as input for more
fine-grained reporting.