session/pset: manage psets with new MPIR_Pset infrastructure
This commits adds a new MPIR_Pset structure and functions to manage
a global list of psets. The two psets defined by the MPI standard
mpi://SELF and mpi://WORLD are always in this list. With the added
infrastructure additional psets may be added to the list and existing
psets maybe invalidated.

For now, a MPIR_Pset_member is an int which is assumed to be a rank.
In the future, this data type/structure may need to be extended to also
consider a "world id" - depending on how spawning with sessions will
move forward.