mpir_pmix: add PMIx event handlers for pset definition and deletion
PMIx Psets can be defined and deleted at runtime by PMIx servers.
Every time this happens, a PMIx event is emitted by the PMIx server.
These events are treated by the PMIx event handlers added with this
commit. They add new psets to the global pset list and invalidate
psets that are deleted by the server.

PMIx Psets are available as of PMIx standard version 4, OpenPMIx
major version 4. A macro is introduced to check for OpenPMIx
min version. This helps to avoid a clash with MPICH's PMIx implementation
where the version variables defined by OpenPMIx are not available.