mpir_pmix: add basic infrastructure for pmix proc id to mpi rank mapping.
This commit adds a basic set of functions and objects to manage the
mapping of PMIx process id (nspace, rank) to MPI rank in a generic
way. The approach is based on a global array (pmix_data_array_t) of
all processes that is used to map PMIx procs to MPI ranks and vice
versa.

Currently we are not yet able to detect multiple namespaces in
pmix_init. There is future work required to provide a full support
for this using PMIx Process Group and PMIx Connect mechanisms.
The changes of this commit serve as a basis for this to enable
dynamic process management (for MPI Sessions) in the future.