binding/python: refactor binding_c.py
Refactor code to function dump_body_of_routine.

This is a preparation for generating io bindings. Since ROMIO is
designed on top of the rest of MPI, we will handle differently for
parameter filtering and error checking, but the part that handles large
count and calling the internal impl functions can be shared.