ch3/ch4: add MPID_Pset_member_get_lpid to map pset member to lpid
As the lpids are managed on device level, a mapping function is required
to get from MPIR_Pset_member to lpid. For now, this mapping is trivial
but this might change in the future depending on how MPIR_Pset_member type
and spawning plus MPI sessions evolves in the future.