ADI/comm: exchange context_id in MPID_Intercomm_exchange
We can easily excahgne the context_id along with the rest of the remote
info rather than do it in a separate step.

Rename MPID_Intercomm_exchange_map to MPID_Intercomm_exchange to better
reflect that it is not just exchanging maps.

Replace the type for is_low_group from int to bool.