ch3: remove the isDisconnected arg from MPIDI_VCRT_Release
The only logic for whether to release a vc is whether this vc is for a
dynamic process. It has nothing to do with the whether
MPI_Comm_disconnect is called. The semantics of MPI_Comm_disconnect is
just to wait for all communication complete. It is orthogonal to how the
comm is destroyed.