ch4/ofi: add a way to tell whether an av entry is empty
When ch4-layer allocates an av table, all entries are initialized to 0.
However, 0 can be a valid entry for fi_addr_t. We could initialize all
entries to FI_ADDR_NOTAVAIL, but that requires an additional complexity
of a netmod API. Instead, because the entry 0 is always the first entry
to be inserted by fi_av_insert, we can simply remember the entry
(MPIDI_OFI_global.lpid0) and be able to tell which entries are empty (in
MPIDI_OFI_insert_upid).