ch4/ofi: do vci address exchange within a comm
Add the flexibility of perform multiple vci address exchange within a
comm other than the comm world. For one, this is important in a session
where the comm_world may not exist. For two, this provides a mechanism
to save resource when applications don't need multiple vcis for the
entire comm world.