pmi: do not add src/pmi/include/pmix to CPPFLAGS
We put embedded pmix.h to the pmix subdir to avoid clashing with system
pmix.h. Adding it to CPPFLAGS makes it depending on the rather obscure
inclusion order. Rather, include pmix/pmix.h explicitly in the code when
we know we are building embedded pmix.

This is consistent with how we do it in src/pmi/src/pmix.c.