mpid/shm: remove MPIDU_shm_seg_t and refactor
MPIDU_shm_seg_t was used by mpidu_shm_alloc.c, mpidu_init_shm.c, and
mpidu_init_shm_alloc.c. However, the usages are all slightly different
and some fields are only used in one but not the other. It is simpler to
locally define it or, in the case of mpidu_init_shm.c, just use static
globals.