datatype: allow MPI_Type_size on unsupported types
Builitin MPI types may not be supported, for example, when the compiler
does not support the type. A common case is for Fortran types such as
MPI_REAL when MPICH is configured without Fortran binding. Since there
is no standard way for users check whether a builtin type is supported,
MPI_Type_size can be used as a check. MPI_Type_size will return 0 if a
builtin type is not supported. For MPI_Type_size to work as a check, it
should not return error when the datatype is not supported. Thus, we
need a custom error_check, rather than the default MPIR_ERRTEST_DATATYPE
which will return error  on unsupported types.