ch4/ofi: clean up MPIDI_OFI_AV_ADDR macros
The MPIDI_OFI_AV_ADDR macros are difficult to read. It is only used in
MPIDI_OFI_av_to_phys and addr_exchange.

Expand the macro in both places so it can use if/else and add
proper comments.

Also replace the redundant macro MPIDI_OFI_AV_ROOT_ADDR, which is the
same as MPIDI_OFI_AV_ADDR_ROOT.