ch4/xpmem: Support handle unmapping
When the mapped segment cache is disabled, unmap segments once we are
done copying from them. Since we need access to the actual attached
address, we store it in the handle during mapping. To achieve this, we
pass the handle to map/unmap by reference rather than value to allow
modification of its contents.