typerep: Call device commit hook when unflattening type
Unflattened types are considered "committed" and used in
communication. When they are eventually freed, the device free hook will
be invoked, so we need to make sure that the commit hook is invoked at
creation time. Otherwise we may read uninitialized values during
free. Issue was found using valgrind and the ch4:ucx device with PETSc
CI tests.

Reported-by: Satish Balay <balay@mcs.anl.gov>