hydra: support group-level PMI_Barrier
PMI_Barrier wire command may have an optional "group=..." attribute,
which restrict the barrier within the process group specified by the
group. The group string include special names such as "WORLD", "NODE",
or a comma-separated list of world ranks. If the group attribute is
missing, it defaults to "WORLD", which is backward compatible to
previous PMI-1 versions (1.0 and 1.1).