hydra/mpiexec: store min_node_id in pg for proxy lookup
In preparation to support group PMI_Barrier, we need a way to map from a
list of global ranks to proxies in a pg. The rankmap maps from global
rank to node_id, and we need a min_node_id offset to map from node_id to
the index in pg->proxy_list.