comm: add comm->attr
When comm->attr is 0, many fields in MPIR_Comm are unset. The comm
should still work but all the optimization parts should check comm->attr
before accessing the corresponding comm->fields.

This allows for lightweight internal comm that can be used in boot
strapping or in commposite algorithms that do not need complex and heavy
full Comm_commit.