comm: add MPIR_COMM_ATTR__SUBCOMM
There is no easy way to tell whether a comm is a user-visible or an
internal subcomm. Add a attr bit for it.

NOTE: subcomms are only used in collectives. If we swap the ranks in
MPIC send/recv functions and convert the subcomm traffic into parent
comm, then we don't need context offset such as
MPIR_CONTEXT_INTRANODE_OFFSET and MPIR_CONTEXT_INTERNODE_OFFSET.
In addition, we can skip most of facilities in the subcomm achieving the
same goals as in https://github.com/pmodels/mpich/pull/7103.