ch4: remove MPIDI_global.is_initialized
Since we are going to remove world init, there won't be a global
is_initialized flag. This flag is used to signal ofi_send to use a
simpler and more robust send_fallback during init and bootstrapping
stage. This is still desirable, but we need figure out a way to pass the
calling context (init or bootstrapping) down to the netmod send.