ch4/posix: move allgather of topo to comm_commit_hook
Move allgather of topo info to MPIDI_POSIX_mpi_comm_commit_post_hook
since we won't have a world init stage. For now, we only do it if we are
initializing a comm_world. It is possible to do this in comm_bootstrap,
but for now this will do. It just means the topo aware optimization
won't be available in a session model.