comm: reserve MPIR_CTXID_BOOTSTRAP
We need a reserved context id in order to run
MPIR_Get_contextid_sparse_group. This in itself isn't sufficient because
the groups may overlap and the group ranks are not unique between groups.
For now, we'll rely on a user supplied tag to ensure isolation.

Move the macro definition to mpir_contextid.h so we can use it in
MPIR_Comm_create_from_group_impl in comm_impl.c. Because
mpir_contextid.h comes before mpidpre.h in which MPID_NEEDS_ICOMM_WORLD
is defined, we reserves MPIR_CTXID_ICOMM_WORLD regardless of the option.
I think this is a small price for cleaner code.

NOTE: in theory, if we always use global rank to run
MPIR_Get_contextid_sparse_group then tag won't be needed to prevent
mismatch.