mpir/pmix: picking PMIX_PREFIX fix from psmpi
PMIx client need to remove PMIX_PREFIX info key or the server may
complain on inconsistent number of info items, e.g. return
PMIX_ERR_EMPTY. This is fixed in OpenPMIx 5.0.7. But in-between the
upgrade, we need this patch to omit setting PMIX_PREFIX key.

Thanks Sonja Happ for the fix!

Co-authored-by: Sonja Happ <sonja.happ@par-tec.com>