all: remove icomm_world
Now that we no longer have a special world init stage, there is no need
for the MPIR layer to initialize and finalize the potential internal
icomm_world. If the device layer needs the internal comm, it can simply
use dup the comm_world at the end of MPID_Comm_commit_pre_hook or
MPID_Comm_commit_post_hook and maintain it inside the device layer.