pmi: do not make exitcode optional in PMI Abort
There is little merit making exitcode optional in the PMI wire protocol.

In commit 60a5486d7, we skip optional parameters altogether which made
the exitcode disappear. This broke gforker. We certainly could make
gforker to be more robust by assuming a default. But such optional
behavior adds the complexity to PMI-1 protocol. Thus, we choose to keep
it simple here.