mpl/shm: Fix race in MPL_initshm_open
The creation of the shared memory fd by shm_open is atomic with respect
to other processes, but the region is initially of size 0. To avoid
trying to mmap the region before it has been resized by the root
process, check the size via fstat until it matches the expected value.