ch4: move MPIR_nodeid_init into MPID_InitCompleted
The MPIR_nodeid_init is meant to support shared memory for dynamic processes.
It is only needed by supported device and it is only used in the world
model. In particular, it assumes usage of MPIR_Process.comm_world. Move
the call to MPID_InitCompleted to give the device more control.