doc: Remove RELEASE_NOTES
The content of RELEASE_NOTES is out of date and not especially
useful. Many of the restrictions or limitations are either encoded into
the configure script, or can be found in open issues on the MPICH GitHub
page. Point people to GitHub from the README for the most current
information.